ArticlesReccomender/requirements.txt
pandas~=1.3.2
elsapy~=0.5.0
tqdm~=4.42.1
habanero~=0.7.4
beautifulsoup4
Django
djangorestframework
markdown
django-filter
numpy
scikit-surprise
